Compress profile files with gzip
This can reduce profile file size by multiple orders of magnitude. In my
benchmarks I found the compressed files to be up to 97% (!) smaller than
uncompressed ones.  The main reason for this is that traces contain the
same instruction pointers again and again; thus they are very well
suited for compression.

Old (uncompressed) profile files are still supported. In fact, the
profile file format did not change, it only got a gzip wrapper. In the
reader we simply check if a given profile is gzip-compressed using the
gzip magic bytes "0x1f 0x8b"; if it is compressed, we decompress it and
then continue parsing the now-uncompressed profile just like we parse an
uncompressed profile.


Former-commit-id: e8c917618a7f58b16432dc02d6e439f5137b0ce9